# swagger.yml
openapi: 3.0.3
info:
  title: "NUPEVID API"
  description: "API do sistema NUPEVID para gerenciamento de vítimas, atendimentos e medidas protetivas"
  version: "1.0.0"
  contact:
    name: "Equipe NUPEVID"
servers:
  - url: "/api/v1"
    description: "Servidor Principal"
...
security:
  - bearerAuth: []
  - api_key: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "Autenticação via Token JWT obtido no endpoint /auth/login"
    api_key:
      type: apiKey
      in: header
      name: api_key
      description: "Chave de API configurada nas variáveis de ambiente"
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: string
          description: "Tipo de erro"
          example: "Bad Request"
        message:
          type: string
          description: "Mensagem de erro descritiva"
          example: "Validation error: field X is required"
        status_code:
          type: integer
          format: int32
          description: "Código de status HTTP"
          example: 400
...
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        rank:
          type: string
          example: "CAP PM"
        registration:
          type: string
          example: "100000001"
        full_name:
          type: string
          example: "João Silva"
        profile:
          type: string
          enum: ["ROOT", "CITY_ADMIN", "CITY_USER"]
        email:
          type: string
          format: email
        city_id:
          type: string
          format: uuid
          nullable: true
        permission_policies:
          type: object
          description: "Políticas de permissão do usuário"
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        is_deleted:
          type: boolean
...
    CreateUserRequest:
      type: object
      required:
        - rank
        - registration
        - full_name
        - profile
        - email
        - password
      properties:
        rank:
          type: string
          example: "CB PM"
        registration:
          type: string
          pattern: "^1000[0-9]{5}$"
          example: "100000001"
        full_name:
          type: string
          example: "Maria Santos"
        profile:
          type: string
          enum: ["ROOT", "CITY_ADMIN", "CITY_USER"]
        email:
          type: string
          format: email
        password:
          type: string
          format: password
          minLength: 8
        city_id:
          type: string
          format: uuid
          nullable: true
        permission_policies:
          type: object
          description: "Políticas de permissão opcionais"
    UpdateUserRequest:
      type: object
      required:
        - rank
        - registration
        - full_name
        - profile
        - email
      properties:
        rank:
          type: string
        registration:
          type: string
        full_name:
          type: string
        profile:
          type: string
          enum: ["ROOT", "CITY_ADMIN", "CITY_USER"]
        email:
          type: string
          format: email
        city_id:
          type: string
          format: uuid
          nullable: true
        permission_policies:
          type: object
    UpdatePasswordRequest:
      type: object
      required:
        - current_password
        - new_password
      properties:
        current_password:
          type: string
          format: password
          description: "Senha atual (obrigatória)"
        new_password:
          type: string
          format: password
          minLength: 8
          description: "Nova senha (obrigatória para todos)"
...
paths:
...
  /users:
    get:
      tags:
        - "Users"
      summary: "Listar todos os usuários"
      description: "Retorna lista de usuários com filtros baseados nas permissões do usuário autenticado"
      operationId: "getAllUsers"
      security:
        - bearerAuth: []
          api_key: []
      parameters:
        - name: "page"
          in: "query"
          required: false
          schema:
            type: integer
            example: 1
          description: "Número da página (1-based)"
        - name: "page_size"
          in: "query"
          required: false
          schema:
            type: integer
            example: 10
          description: "Quantidade de itens por página (max 200)"
      responses:
        "200":
          description: "Lista de usuários retornada com sucesso"
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: integer
                    example: 200
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/User"
                  page:
                    type: integer
                    example: 1
                  page_size:
                    type: integer
                    example: 10
                  total_items:
                    type: integer
                    example: 100
                  total_pages:
                    type: integer
                    example: 10
              example:
                message: "Operation successful"
                status: 200
                data:
                  - id: "550e8400-e29b-41d4-a716-446655440000"
                    rank: "CAP PM"
                    registration: "100000001"
                    full_name: "João Silva"
                    profile: "CITY_ADMIN"
                    email: "joao.silva@pm.ro.gov.br"
                    city_id: "660e8400-e29b-41d4-a716-446655440000"
                    permission_policies: {}
                    created_at: "2025-01-15T10:30:00Z"
                    updated_at: "2025-01-15T10:30:00Z"
                    is_deleted: false
                page: 1
                page_size: 10
                total_items: 1
                total_pages: 1
        "401":
          description: "Token JWT ausente, inválido ou expirado"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error: "Unauthorized"
                message: "Unauthorized"
                status_code: 401
        "403":
          description: "Sem permissão para listar usuários"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error: "Forbidden"
                message: "Forbidden: insufficient permissions"
                status_code: 403

    post:
      tags:
        - "Users"
      summary: "Criar novo usuário"
      description: "Cria um novo usuário no sistema. Apenas ROOT e CITY_ADMIN podem criar usuários"
      operationId: "createUser"
      security:
        - bearerAuth: []
          api_key: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserRequest"
      responses:
        "201":
          description: "Usuário criado com sucesso"
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: integer
                    example: 201
                  data:
                    $ref: "#/components/schemas/User"
              example:
                message: "Operation successful"
                status: 201
                data:
                  id: "550e8400-e29b-41d4-a716-446655440000"
                  rank: "SD PM"
                  registration: "100000002"
                  full_name: "Maria Souza"
                  profile: "CITY_USER"
                  email: "maria.souza@pm.ro.gov.br"
                  city_id: "660e8400-e29b-41d4-a716-446655440000"
                  permission_policies: {}
                  created_at: "2025-01-15T10:30:00Z"
                  updated_at: "2025-01-15T10:30:00Z"
                  is_deleted: false
        "400":
          description: "Dados inválidos"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error: "Bad Request"
                message: "Bad Request: city_id is required for CITY_ADMIN and CITY_USER profiles"
                status_code: 400
        "401":
          description: "Token JWT ausente, inválido ou expirado"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error: "Unauthorized"
                message: "Unauthorized"
                status_code: 401
        "403":
          description: "Sem permissão para criar usuários"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error: "Forbidden"
                message: "Forbidden: insufficient permissions"
                status_code: 403
...
  /users/{id}:
    parameters:
      - name: "id"
        in: "path"
        required: true
        schema:
          type: string
          format: uuid
        description: "ID do usuário"

    get:
      tags:
        - "Users"
      summary: "Obter usuário por ID"
      description: "Retorna detalhes de um usuário específico"
      operationId: "getUserById"
      security:
        - bearerAuth: []
          api_key: []
      responses:
        "200":
          description: "Usuário encontrado"
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: integer
                    example: 200
                  data:
                    $ref: "#/components/schemas/User"
              example:
                message: "Operation successful"
                status: 200
                data:
                  id: "550e8400-e29b-41d4-a716-446655440000"
                  rank: "CAP PM"
                  registration: "100000001"
                  full_name: "João Silva"
                  profile: "CITY_ADMIN"
                  email: "joao.silva@pm.ro.gov.br"
                  city_id: "660e8400-e29b-41d4-a716-446655440000"
                  permission_policies: {}
                  created_at: "2025-01-15T10:30:00Z"
                  updated_at: "2025-01-15T10:30:00Z"
                  is_deleted: false
        "401":
          description: "Token JWT ausente, inválido ou expirado"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error: "Unauthorized"
                message: "Unauthorized"
                status_code: 401
        "403":
          description: "Sem permissão para visualizar este usuário"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error: "Forbidden"
                message: "Forbidden: insufficient permissions"
                status_code: 403
        "404":
          description: "Usuário não encontrado"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error: "Not Found"
                message: "Not Found: user not found"
                status_code: 404

    put:
      tags:
        - "Users"
      summary: "Atualizar usuário"
      description: "Atualiza informações de um usuário existente"
      operationId: "updateUser"
      security:
        - bearerAuth: []
          api_key: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserRequest"
      responses:
        "200":
          description: "Usuário atualizado com sucesso"
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: integer
                    example: 200
                  data:
                    $ref: "#/components/schemas/User"
              example:
                message: "Operation successful"
                status: 200
                data:
                  id: "550e8400-e29b-41d4-a716-446655440000"
                  rank: "MAJ PM"
                  registration: "100000001"
                  full_name: "João Silva"
                  profile: "CITY_ADMIN"
                  email: "joao.silva@pm.ro.gov.br"
                  city_id: "660e8400-e29b-41d4-a716-446655440000"
                  permission_policies: {}
                  created_at: "2025-01-15T10:30:00Z"
                  updated_at: "2025-02-01T14:00:00Z"
                  is_deleted: false
        "400":
          description: "Dados inválidos"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error: "Bad Request"
                message: "Bad Request: CITY_ADMIN already exists for this city"
                status_code: 400
        "401":
          description: "Token JWT ausente, inválido ou expirado"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error: "Unauthorized"
                message: "Unauthorized"
                status_code: 401
        "403":
          description: "Sem permissão para atualizar este usuário"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error: "Forbidden"
                message: "Forbidden: insufficient permissions"
                status_code: 403
        "404":
          description: "Usuário não encontrado"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error: "Not Found"
                message: "Not Found: user not found"
                status_code: 404

    delete:
      tags:
        - "Users"
      summary: "Deletar usuário (soft delete)"
      description: "Realiza soft delete de um usuário"
      operationId: "deleteUser"
      security:
        - bearerAuth: []
          api_key: []
      responses:
        "200":
          description: "Usuário deletado com sucesso"
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: integer
                    example: 200
                  data:
                    $ref: "#/components/schemas/User"
              example:
                message: "Operation successful"
                status: 200
                data:
                  id: "550e8400-e29b-41d4-a716-446655440000"
                  rank: "CAP PM"
                  registration: "100000001"
                  full_name: "João Silva"
                  profile: "CITY_ADMIN"
                  email: "joao.silva@pm.ro.gov.br"
                  city_id: "660e8400-e29b-41d4-a716-446655440000"
                  permission_policies: {}
                  created_at: "2025-01-15T10:30:00Z"
                  updated_at: "2025-02-01T14:00:00Z"
                  is_deleted: true
        "401":
          description: "Token JWT ausente, inválido ou expirado"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error: "Unauthorized"
                message: "Unauthorized"
                status_code: 401
        "403":
          description: "Sem permissão para deletar este usuário"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error: "Forbidden"
                message: "Forbidden: insufficient permissions"
                status_code: 403
        "404":
          description: "Usuário não encontrado"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error: "Not Found"
                message: "Not Found: user not found"
                status_code: 404
  /users/password:
    patch:
      tags:
        - "Users"
      summary: "Atualizar senha do usuário atual"
      description: |
        Atualiza a senha do usuário autenticado.

        **Regras:**
        - É obrigatório informar a senha atual
      operationId: "updateUserPassword"
      security:
        - bearerAuth: []
          api_key: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePasswordRequest"
            examples:
              usuario_comum:
                summary: "Usuário alterando sua própria senha"
                value:
                  current_password: "SenhaAtual123!"
                  new_password: "NovaSenha456!"
      responses:
        "200":
          description: "Senha atualizada com sucesso"
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: integer
                    example: 200
                  data:
                    $ref: "#/components/schemas/User"
              example:
                message: "Operation successful"
                status: 200
                data:
                  id: "550e8400-e29b-41d4-a716-446655440000"
                  rank: "CAP PM"
                  registration: "100000001"
                  full_name: "João Silva"
                  profile: "CITY_ADMIN"
                  email: "joao.silva@pm.ro.gov.br"
                  city_id: "660e8400-e29b-41d4-a716-446655440000"
                  permission_policies: {}
                  created_at: "2025-01-15T10:30:00Z"
                  updated_at: "2025-02-01T14:00:00Z"
                  is_deleted: false
        "400":
          description: "Senha atual incorreta ou dados inválidos"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error: "Bad Request"
                message: "Bad Request: current password is incorrect"
                status_code: 400
        "401":
          description: "Token JWT ausente, inválido ou expirado"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error: "Unauthorized"
                message: "Unauthorized"
                status_code: 401
...